+++ api/CHANGELOG.md
@@ -38,6 +38,7 @@
 - Added a loading screen that appears when selecting an app from storage from the `Select an app` screen https://github.com/Jman-Github/Universal-ReVanced-Manager/issues/255
 - Added support for the ReVanced Patcher v22 while keeping backwards compatibility with v21 https://github.com/Jman-Github/Universal-ReVanced-Manager/issues/252
 - Overall improved the stability to all patcher runtimes
+- Added device architecture, Android version, device model, and patcher version to the patcher logs
 
 
 # Bug fixes
@@ -68,6 +69,8 @@
 - Fixed an issue where the patcher process would die for some users https://github.com/Jman-Github/Universal-ReVanced-Manager/issues/267
 - Fixed an issue where the ReVanced v22 patcher runtime wouldn't follow the experimental patcher setting
 - Fixed an issue where the patching process would die (needs testing) https://github.com/Jman-Github/Universal-ReVanced-Manager/issues/267
+- Fixed issues where the patcher screen would not show the progress of patching accurately
+- Fixed issues where the ReVanced v22 runtime would patch incorrectly resulting in broken patched apps
 
 
 # CI